[DEV-3724] Updating IDV Award Summary Page, Description Section (#1453)
* [DEV-3724] moving psc object traversal
logic to helper & using it in IDV model

* [DEV-3724] removing isIdv conditional logic

* [DEV-3724] fixing styling issues w/ glossary icon
wrapping to the next line by itself

* [DEV-3724] moving to pscHelper namespace

* [DEV-3724] test for helper file

* [DEV-3724] tests for models

* [DEV-3724] fixing lint error

* [DEV-3724] removing second lint error
