+++ src/js/components/award/shared/description/AwardDescription.jsx
@@ -29,7 +29,6 @@ const AwardDescription = ({
     psc = null,
     awardType
 }) => {
-    const isIdv = (awardType === 'idv');
     const tooltip = getToolTipBySectionAndAwardType('description', awardType);
     return (
         <AwardSection type="column" className="award-viz award-description">
@@ -40,28 +39,30 @@ const AwardDescription = ({
                     <div className="award-description__naics-psc">
                         <div className="naics-psc__section">
                             <div className="naics-psc__heading">
-                                North American Industry Classification System (NAICS) Code
-                                <span className="naics-psc__icon">
+                                North American Industry Classification System (NAICS)
+                                <span>
+                                    {/* last word of heading inside the span to prevent the glossary icon from wrapping to its own line by itself */}
+                                    Code
                                     <a href={`#/award/${awardId}/?glossary=naics`}>
                                         <Glossary alt="View glossary definition of NAICS" />
                                     </a>
                                 </span>
                             </div>
-                            {!isIdv && <LineTree type="naics" data={naics} />}
-                            {isIdv && naics}
+                            <LineTree type="naics" data={naics} />
                         </div>
                         <div className="naics-psc__section naics-psc__section_psc">
                             <div className="naics-psc__section">
                                 <div className="naics-psc__heading">
-                                    Product or Service Code (PSC)
-                                    <span className="naics-psc__icon">
+                                    Product or Service Code
+                                    <span>
+                                        {/* last word of heading inside the span to prevent the glossary icon from going to its own line by itself */}
+                                        (PSC)
                                         <a href={`#/award/${awardId}/?glossary=productservice-code-psc`}>
                                             <Glossary alt="View glossary definition of Product/Service Code (PSC)" />
                                         </a>
                                     </span>
                                 </div>
-                                {!isIdv && <LineTree type="psc" data={psc} />}
-                                {isIdv && psc}
+                                <LineTree type="psc" data={psc} />
                             </div>
                         </div>
                     </div>
