+++ tests/models/awardsV2/BaseContract-test.js
@@ -56,27 +56,4 @@ describe('BaseContract', () => {
     it(' should create a parent award details property', () => {
         expect(Object.getPrototypeOf(contract.parentAwardDetails)).toEqual(BaseParentAwardDetails);
     });
-    describe('Deducing the PSC Type based on the PSC Top Tier Code from the API', () => {
-        it.each([
-            ['PRODUCTS', '4'],
-            ['RESEARCH AND DEVELOPMENT', 'A'],
-            ['SERVICES', 'B'],
-            ['PRODUCTS', undefined]
-        ])(
-            ('psc.toptier_code.description should be %s when %s is the psc.toptier_code.code'),
-            (result, pscCode) => {
-                const mockData = {
-                    ...mockContract,
-                    psc_hierarchy: {
-                        toptier_code: {
-                            code: pscCode,
-                            description: ''
-                        }
-                    }
-                };
-                contract.populate(mockData);
-                expect(contract.psc.pscType.description).toEqual(result);
-            }
-        );
-    });
 });
