+++ tests/models/awardsV2/BaseIdv-test.js
@@ -48,4 +48,27 @@ describe('BaseIdv', () => {
             expect(Object.getPrototypeOf(contract.executiveDetails)).toEqual(CoreExecutiveDetails);
         });
     });
+    describe('Deducing the PSC Type based on the PSC Top Tier Code from the API', () => {
+        it.each([
+            ['PRODUCTS', '4'],
+            ['RESEARCH AND DEVELOPMENT', 'A'],
+            ['SERVICES', 'B'],
+            ['PRODUCTS', undefined]
+        ])(
+            ('psc.toptier_code.description should be %s when %s is the psc.toptier_code.code'),
+            (result, pscCode) => {
+                const mockData = {
+                    ...mockIdv,
+                    psc_hierarchy: {
+                        toptier_code: {
+                            code: pscCode,
+                            description: ''
+                        }
+                    }
+                };
+                contract.populate(mockData);
+                expect(contract.psc.pscType.description).toEqual(result);
+            }
+        );
+    });
 });
